refactor(model-service): flatten layout — drop integrations/, rename sse_utils→sse, merge traj_*→traj
The integrations/ directory only ever held two files (traj_recorder, traj_replayer)
and the litellm CustomLogger angle that justified the name is long gone. Both
modules share one JSONL schema, so collapsing them into a single traj.py keeps
the schema and its read/write halves visible together.

sse_utils.py → sse.py: the codec is the module's whole purpose, the _utils
suffix added nothing.

Drop traj_recorder.now() — a one-line wrapper around time.time() with no
callers.

Also remove a stray _get_or_create_metrics_monitor patch in
test_forward_invokes_recorder_on_success: OTLP create-time failure only logs
a warning, so the patch was protecting against nothing.
